+++ stocks/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Stock(models.Model):
 
    name = models.CharField(max_length=40, unique=True)
-   current_price = models.IntegerField()
+   current_price = models.FloatField()
 
    #num_owned = models.IntegerField(default=0)
    #low = models.IntegerField(default=current_price)
